refactor: refactor save logic in MemberPost
