+++ post/src/main/java/com/oing/controller/MemberPostController.java
@@ -9,12 +9,9 @@ import com.oing.dto.response.PaginationResponse;
 import com.oing.dto.response.PostResponse;
 import com.oing.dto.response.PreSignedUrlResponse;
 import com.oing.exception.AuthorizationFailedException;
-import com.oing.exception.DuplicatePostUploadException;
-import com.oing.exception.InvalidUploadTimeException;
 import com.oing.restapi.MemberPostApi;
 import com.oing.service.MemberBridge;
 import com.oing.service.MemberPostService;
-import com.oing.util.IdentityGenerator;
 import com.oing.util.PreSignedUrlGenerator;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -23,8 +20,6 @@ import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.stereotype.Controller;
 
 import java.time.LocalDate;
-import java.time.LocalTime;
-import java.time.ZonedDateTime;
 
 /**
  * no5ing-server
@@ -37,7 +32,6 @@ import java.time.ZonedDateTime;
 @Controller
 public class MemberPostController implements MemberPostApi {
 
-    private final IdentityGenerator identityGenerator;
     private final PreSignedUrlGenerator preSignedUrlGenerator;
     private final MemberPostService memberPostService;
     private final MemberBridge memberBridge;
@@ -73,46 +67,11 @@ public class MemberPostController implements MemberPostApi {
             key = "#loginFamilyId.concat(':').concat(T(java.time.format.DateTimeFormatter).ofPattern('yyyy-MM').format(#request.uploadTime()))")
     public PostResponse createPost(CreatePostRequest request, String loginFamilyId, String loginMemberId) {
         log.info("Member {} is trying to create post", loginMemberId);
-        String postId = identityGenerator.generateIdentity();
-        ZonedDateTime uploadTime = request.uploadTime();
 
-        validateUserHasNotCreatedPostToday(loginMemberId, loginFamilyId, uploadTime);
-        validateUploadTime(loginMemberId, uploadTime);
-
-        String postImgKey = preSignedUrlGenerator.extractImageKey(request.imageUrl());
-        MemberPost post = new MemberPost(postId, loginMemberId, loginFamilyId, request.imageUrl(),
-                postImgKey, request.content());
-        MemberPost savedPost = memberPostService.save(post);
-        log.info("Member {} has created post {}", loginMemberId, postId);
+        MemberPost savedPost = memberPostService.createMemberPost(request, loginMemberId, loginFamilyId);
+        log.info("Member {} has created post {}", loginMemberId, savedPost.getId());
 
         return PostResponse.from(savedPost);
-    }
-
-    private void validateUserHasNotCreatedPostToday(String memberId, String familyId, ZonedDateTime uploadTime) {
-        LocalDate today = uploadTime.toLocalDate();
-        if (memberPostService.hasUserCreatedPostToday(memberId, familyId, today)) {
-            log.warn("Member {} has already created a post today", memberId);
-            throw new DuplicatePostUploadException();
-        }
-    }
-
-    /**
-     * 업로드 시간이 허용 가능한 범위 내에 있는지 검증합니다.
-     * 범위는 서버의 로컬 시간을 기준으로 전 날의 오후 12시부터 다음 날의 오후 12시까지로 정의됩니다.
-     *
-     * @param uploadTime 검증할 업로드 시간입니다.
-     * @throws InvalidUploadTimeException 업로드 시간이 허용 가능한 범위를 벗어난 경우 발생하는 예외입니다.
-     */
-    private void validateUploadTime(String memberId, ZonedDateTime uploadTime) {
-        ZonedDateTime serverTime = ZonedDateTime.now();
-
-        ZonedDateTime lowerBound = serverTime.minusDays(1).with(LocalTime.of(12, 0));
-        ZonedDateTime upperBound = serverTime.plusDays(1).with(LocalTime.of(12, 0));
-
-        if (uploadTime.isBefore(lowerBound) || uploadTime.isAfter(upperBound)) {
-            log.warn("Member {} is attempting to upload a post at an invalid time", memberId);
-            throw new InvalidUploadTimeException();
-        }
     }
 
     @Override
